backend/requirements.txt
asgiref==3.11.1
coverage==7.13.5
Django==5.1.6
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
gunicorn==23.0.0
packaging==26.0
psycopg2-binary==2.9.10
PyJWT==2.12.1
python-dotenv==1.0.1
sqlparse==0.5.5
tzdata==2026.1
